requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
Django==4.2.3
django-ckeditor==6.6.1
django-cors-headers==4.1.0
django-jazzmin==2.6.0
django-js-asset==2.1.0
django-multiupload==0.6.1
djangorestframework==3.14.0
Pillow==10.0.0
pytz==2023.3
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.7.1
